modules/location_grpc_producer_svc/requirements.txt
grpcio
protobuf
kafka-python